4th

+++ src/kadai3/Calculator.java
@@ -5,20 +5,27 @@ public class Calculator {
 	public static void main(String[] args) {
 		int a = add(3, 7);
 		System.out.println(a);
+		
 		int b = sub(3, 7);
 		System.out.println(b);
+		
 		double c = avg(3, 7);
 		System.out.println(c);
-
 	}
 
-	static int add(int x,int y) {
+	//足し算メソッド
+	static int add(int x, int y) {
 		return x + y;
 	}
-	static int sub(int x,int y) {
+	
+	//引き算メソッド
+	static int sub(int x, int y) {
 		return x - y;
 	}
-	static double avg(double x,double y) {
-		return (x + y )/2;
+	
+	//平均値メソッド（Step2）
+	static double avg(int x, int y) {
+		return (x + y) / 2.0;
+		//return ((double)x + y) / 2;
 	}
 }
